Applied Nav's patch
Summary:
From Naveed:
```
Heya JB

I have to  make a patch for the RNAseq workflow being used by OTP to avoid the seg fault issue in feature counts. I introduced a $FEATURE_COUNT_CORES variable for this. Also, I added a $useSingleEndProcessing that was missing in the configuration XML.

Below is the git diff of my patched version 1.0.22-1 to the previous 1.0.22. Can you add these to the version which you are working on?

Thanks!

Nav
```

Test Plan: .

Reviewers: ishaque

Differential Revision: https://eilslabs-phabricator.dkfz.de/D425