Rolled back the single-quotes around ADAPTER_SEQ in STAR_PARAMS_CLIP.
Without these quotes STAR can identify all adapters as adapters. E.g.

--clip3pAdapterSeq AGATCGGAAGAGCACACGTCTGAACTCCAGTCAC AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAGATCTCGGTGGTCGCCGTATCATT

results in a .Log.out file stating

clip3pAdapterSeq                  AGATCGGAAGAGCACACGTCTGAACTCCAGTCAC   AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAGATCTCGGTGGTCGCCGTATCATT

Note that both values are assigned to the correct parameter.

Compare now to

--clip3pAdapterSeq 'AGATCGGAAGAGCACACGTCTGAACTCCAGTCAC AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAGATCTCGGTGGTCGCCGTATCATT'

which results in a .Log.out file stating

clip3pAdapterSeq                  "AGATCGGAAGAGCACACGTCTGAACTCCAGTCAC AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAGATCTCGGTGGTCGCCGTATCATT"

Note that the single-quotes are substituted by double quotes. STAR seems to group the two together. It is not clear what STAR does with an adapter sequence containing a space, but certainly not an error message or even a warning.

As the STAR manual never uses quotes also for space-delimited argument lists to parameters the first, unquoted version is conformant to the documentation.